requirements.txt
flask==2.3.2
flask-pymongo==2.3.0
pymongo==4.3.3
langchain==0.0.228
openai==0.27.8
python-dotenv==1.0.0
google-search-results==2.4.2
tiktoken==0.3.0
werkzeug==2.3.4